+++ src/Oro/Bundle/ImapBundle/Manager/ImapClearManager.php
@@ -91,38 +91,33 @@ class ImapClearManager implements LoggerAwareInterface
     /**
      * @param UserEmailOrigin $origin
      */
-    protected function clearOrigin(UserEmailOrigin $origin)
+    protected function clearOrigin($origin)
     {
         $folders = $origin->getFolders();
         $folderRepository = $this->em->getRepository('OroImapBundle:ImapEmailFolder');
 
         foreach ($folders as $folder) {
             $imapFolder = $folderRepository->findOneBy(['folder' => $folder]);
-
-            if (!$origin->isActive()) {
-                $this->removeFolder($imapFolder);
-            } elseif (!$folder->isSyncEnabled()) {
+            if ($imapFolder && !$origin->isActive()) {
+                $this->clearFolder($imapFolder);
+                $this->em->remove($imapFolder);
+                $this->logger->notice(sprintf('ImapFolder with ID %s removed', $imapFolder->getId()));
+            } elseif ($imapFolder && !$folder->isSyncEnabled()) {
                 $this->clearFolder($imapFolder);
                 $imapFolder->getFolder()->setSynchronizedAt(null);
+            }
+        }
+        foreach ($folders as $folder) {
+            if (!$origin->isActive()) {
+                $this->em->remove($folder);
+                $this->logger->notice(sprintf('Folder with ID %s removed', $folder->getId()));
             }
         }
 
         if (!$origin->isActive()) {
             $this->em->remove($origin);
-            $this->em->flush();
+            $this->logger->notice(sprintf('Origin with ID %s removed', $origin->getId()));
         }
-    }
-
-    /**
-     * @param ImapEmailFolder $imapFolder
-     */
-    protected function removeFolder(ImapEmailFolder $imapFolder)
-    {
-        $this->clearFolder($imapFolder);
-
-        $folder = $imapFolder->getFolder();
-        $this->em->remove($imapFolder);
-        $this->em->remove($folder);
 
         $this->em->flush();
     }
@@ -130,41 +125,72 @@ class ImapClearManager implements LoggerAwareInterface
     /**
      * @param ImapEmailFolder $imapFolder
      */
-    protected function clearFolder(ImapEmailFolder $imapFolder)
+    protected function clearFolder($imapFolder)
     {
         $folder = $imapFolder->getFolder();
-
-        $q = $this->em->createQueryBuilder()
-            ->select('eu')
-            ->from('OroEmailBundle:EmailUser', 'eu')
-            ->andWhere('eu.folder = :folder')
-            ->setParameter('folder', $folder)
-            ->getQuery();
-        $iterableResult = $q->iterate();
-
+        $limit = self::BATCH_SIZE;
+        $offset = 0;
         $i = 0;
-        while (($row = $iterableResult->next()) !== false) {
-            /** @var EmailUser $emailUser */
-            $emailUser = $row[0];
-            $email = $emailUser->getEmail();
-            $this->em->remove($emailUser);
+        while ($result =
+            $this->em->getRepository('OroEmailBundle:EmailUser')
+                ->getEmailUserByFolder($folder, $limit, $offset)->getQuery()->getResult()
+        ) {
+            foreach ($result as $emailUser) {
+                /** @var EmailUser $emailUser */
+                $emailUser->removeFolder($folder);
+                $email = $emailUser->getEmail();
+                if ($emailUser->getFolders()->isEmpty()) {
+                    $this->em->remove($emailUser);
+                }
 
-            $imapEmail = $this->em->getRepository('OroImapBundle:ImapEmail')->findOneBy([
-                'email' => $email,
-                'imapFolder' => $imapFolder,
-            ]);
-            if ($imapEmail !== null) {
-                $this->em->remove($imapEmail);
+                $imapEmails = $this->em->getRepository('OroImapBundle:ImapEmail')->findBy([
+                    'email' => $email,
+                    'imapFolder' => $imapFolder
+                ]);
+                foreach ($imapEmails as $imapEmail) {
+                    $this->em->remove($imapEmail);
+                }
+                ++$i;
             }
+            $this->em->flush();
+            $this->cleanUp();
 
-            if (($i % self::BATCH_SIZE) === 0) {
-                $this->em->flush();
-                $this->em->clear('OroEmailBundle:EmailUser');
-                $this->em->clear('OroImapBundle:ImapEmail');
-            }
-            ++$i;
+        }
+        if ($i > 0) {
+            $this->logger->notice(
+                sprintf(
+                    'ImapFolder with ID %s cleared. Removed %d emails.',
+                    $imapFolder->getId(),
+                    $i
+                )
+            );
         }
 
         $this->em->flush();
+        $this->cleanUp();
+    }
+
+    /**
+     * @return array
+     */
+    protected function entitiesToClear()
+    {
+        return [
+            'Oro\Bundle\EmailBundle\Entity\EmailUser',
+            'Oro\Bundle\EmailBundle\Entity\Email',
+            'Oro\Bundle\EmailBundle\Entity\EmailRecipient',
+            'Oro\Bundle\ImapBundle\Entity\ImapEmail',
+            'Oro\Bundle\EmailBundle\Entity\EmailBody',
+        ];
+    }
+
+    /**
+     * clean up
+     */
+    protected function cleanUp()
+    {
+        foreach ($this->entitiesToClear() as $entityClass) {
+            $this->em->clear($entityClass);
+        }
     }
 }
